requirements.txt
feedparser
openai
python-dotenv
flask
mysql-connector-python
